Removed Old Layout + cleanup
