--- app/src/main/res/drawable/colorgrade_base.xml
@@ -1,8 +1,0 @@
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle"
-    android:tint="@color/blue">
-    <solid />
-    <corners
-        android:radius="30dp" />
-
-</shape>
