--- app/src/main/res/drawable/colorgrade_stroke.xml
@@ -1,12 +1,0 @@
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle"
-    android:tint="@color/blue">
-
-    <stroke
-        android:width="5dp"
-        android:dashGap="10dp" />
-    <corners android:radius="30dp" />
-    <solid android:color="#00000000" />
-
-
-</shape>
