+++ app/src/main/res/layout/dialog_symbol_add.xml
