+++ app/src/main/res/layout/table_entry_item.xml
